Further refactoring. Redacted all old code.

+++ src/ff_debug.h
@@ -36,7 +36,7 @@
 #ifdef DEBUG
 
 void DebugLog(const char* log_message);
-void DebugLog(uint8_t source, uint8_t msg_type, uint8_t msg_str, int i_val, float f_val);
+void DebugLog(uint16_t source, uint8_t msg_type, uint8_t msg_str, int i_val, float f_val);
 
 #endif //DEBUG
